rename: request.js파일 main.js로 변경
